setups

+++ scripts/install-functions.sh
@@ -147,14 +147,18 @@ function installDeleteProject() {
 
 function installApigeeOrg() {
   cd "$WORK_DIR"/terraform-modules/apigee-install
-  rm -Rf .terraform*
-  rm terraform.tfstate
+
+  last_project_id=$(cat install-state.txt)
+  if [ -z "$last_project_id" ] || [ last_project_id != "$PROJECT_ID" ]; then
+    rm -Rf .terraform*
+    rm terraform.tfstate
+  fi
   terraform init
   terraform plan -var "apigee_org_create=true" \
-  -var "project_id=$PROJECT_ID" --var-file="$WORK_DIR/terraform-modules/apigee-install/apigee.tfvars"
+    -var "project_id=$PROJECT_ID" --var-file="$WORK_DIR/terraform-modules/apigee-install/apigee.tfvars"
   terraform apply -auto-approve -var "apigee_org_create=true" \
-  -var "project_id=$PROJECT_ID" --var-file="$WORK_DIR/terraform-modules/apigee-install/apigee.tfvars"
-
+    -var "project_id=$PROJECT_ID" --var-file="$WORK_DIR/terraform-modules/apigee-install/apigee.tfvars"
+  "$(terraform output -raw project_id)" > install-state.txt
 }
 
 function fetchHybridInstall() {
